move playlist out of components - only we are using it
