+++ frontend/src/scenes/session-recordings/playlist/Playlist.scss
