+++ frontend/src/scenes/session-recordings/playlist/Playlist.stories.tsx
